Store RPG code scraps on separate branch
After not working on this project for a couple of months, I have
decided to stop development on this project. This is mainly due to
real life responsiblities, my dissatisfaction with the
current code (I feel that I would need to start anew to
properly structure it better), and my desire to switch to a proper
video game engine (LibGDX).

As a result, I decided to group all of the work that I made on the
RPG elements of the game.